flask-app/app/requirements.txt
flask-cors==4.0.0
flask-migrate==4.0.5
flask-script==2.0.6
flask-sqlalchemy==3.1.1
flask==3.0.0
sqlalchemy-utils==0.41.1
sqlalchemy==2.0.21
cryptography==3.4.8
psycopg2-binary==2.9.9
requests==2.28.2
flask-caching==2.1.0
PyJWT==2.8.0